Created a new sermon API on the backend and reprogrammed the front end to handle changes from the API

+++ src/utils/utils.js
@@ -30,7 +30,7 @@ export const nextSundaysDate = () => {
 export const formatDate = (dateString) => {
   const d = new Date(dateString);
   const ye = new Intl.DateTimeFormat("en", { year: "numeric" }).format(d);
-  const mo = new Intl.DateTimeFormat("en", { month: "long" }).format(d);
+  const mo = new Intl.DateTimeFormat("en", { month: "short" }).format(d);
   const da = new Intl.DateTimeFormat("en", { day: "2-digit" }).format(d);
   const formattedDate = `${mo} ${da}, ${ye}`;
   return formattedDate;
